Annotate Vessels Problems
Add commented notes about problematic IF statements in those verbs on the
`liquid` class that check if the liquid's vessel is open:

* examine
* look_in
* pour_in
* put_in

The problem here is that if the vessel is a `listed_container` than the
condition can never be met because the Library caters to ensure that
every closed listed container is set as OPAQUE -- therefore its contents
will be out of scope to the player.

Since the Library does not propose other types of containers, authors
are on their own to implement openable containers which might not be
OPAQUE when closed (i.e. transparent containers). The whole containers
system should be revised to allow a more flexible approach to containers
objects, including transparent containers, and abstract away from end
users opacity management of open/closed containers.